requirements.txt
numpy==1.23.4
Pillow==9.3.0
Pillow==9.5.0
Requests==2.29.0
streamlit==1.22.0
torch==2.0.0
transformers==4.28.1
